Match-id-f0f5ee182c35b9af44dcf63d010de1bcf992e989

+++ build/test.sh
@@ -24,8 +24,12 @@ function prepare(){
 }
 
 function clean() {
-  rm -rf "${TMP_DIR}"
-  rm -rf "${TEST_DIR}"
+  if [ -f "${TMP_DIR}" ]; then
+    rm -rf "${TMP_DIR}"
+  fi
+  if [ -f "${TEST_DIR}" ]; then
+    rm -rf "${TEST_DIR}"
+  fi
   mkdir -p "${TMP_DIR}"
   mkdir -p "${TEST_DIR}"
 }
